// apps/chataigne/ui/src/lib/modules/audio/sound-card/generated/AudioDeviceTargetId.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AudioDeviceId } from "./AudioDeviceId";
import type { BackendId } from "./BackendId";

export type AudioDeviceTargetId = { "kind": "system_default", backend: BackendId, } | { "kind": "device", backend: BackendId, device: AudioDeviceId, };